fix errors, seperate tests
