+++ contracts/test/ERC721A.sol
@@ -233,7 +233,7 @@ contract ERC721A is
   /**
    * @dev Base URI for computing {tokenURI}. If set, the resulting URI for each
    * token will be the concatenation of the `baseURI` and the `tokenId`. Empty
-   * by default, can be overriden in child contracts.
+   * by default, can be overridden in child contracts.
    */
   function _baseURI() internal view virtual returns (string memory) {
     return "";
