+++ package.json
@@ -33,7 +33,7 @@
     "csv-parse": "^5.1.0"
   },
   "devDependencies": {
-    "@dimensiondev/eslint-plugin": "^0.0.2",
+    "@dimensiondev/eslint-plugin": "^0.0.2-20220516081411-2649814",
     "@nomiclabs/hardhat-ethers": "^2.0.6",
     "@nomiclabs/hardhat-solhint": "^2.0.1",
     "@nomiclabs/hardhat-waffle": "^2.0.3",
