+++ project-words.txt
@@ -1,0 +1,29 @@
+arbitrum
+arrayify
+astar
+bignumber
+blockscout
+boba
+buidler
+celo
+chainlink
+Dapplet
+endregion
+eqeqeq
+espace
+hardfork
+hexlify
+kardia
+moonriver
+nomiclabs
+prettierignore
+solhint
+typechain
+xdai
+mockdate
+merkle
+ANFT
+keccak
+IERC
+IQLF
+addrs
