--- test/TestMysteryBox.ts
@@ -1,1463 +1,0 @@
-import { ethers, waffle, upgrades } from "hardhat";
-import { Signer, Contract, utils } from "ethers";
-import MockDate from "mockdate";
-import { use } from "chai";
-import chaiAsPromised from "chai-as-promised";
-const { expect } = use(chaiAsPromised);
-const { deployContract } = waffle;
-import proofs from "../dist/proofs.json";
-
-import { advanceBlock, advanceTimeAndBlock, takeSnapshot, revertToSnapShot } from "./helper";
-import {
-  MaskNFTInitParameters,
-  CreateBoxParameters,
-  openBoxParameters,
-  MaxNumberOfNFT,
-  seconds_in_a_day,
-  holderMinAmount,
-} from "./constants";
-
-const network = "mainnet";
-const maskNftPara = MaskNFTInitParameters[network];
-const createBoxPara = CreateBoxParameters[network];
-let sell_all_box_id: number;
-let not_sell_all_box_id: number;
-let not_sell_all_nft_id_list: number[] = [];
-
-const txParameters = {
-  gasLimit: 6000000,
-  value: createBoxPara.payment[0][1],
-};
-
-import jsonABI from "../artifacts/contracts/MysteryBox.sol/MysteryBox.json";
-const interfaceABI = new ethers.utils.Interface(jsonABI.abi);
-import EnumerableNftTokenABI from "../artifacts/contracts/test/MaskEnumerableNFT.sol/MaskEnumerableNFT.json";
-const EnumerableNftInterface = new ethers.utils.Interface(EnumerableNftTokenABI.abi);
-import nonEnumerableNftTokenABI from "../artifacts/contracts/test/MaskNonEnumerableNFT.sol/MaskNonEnumerableNFT.json";
-
-import TestTokenArtifact from "../artifacts/contracts/test/test_token.sol/TestToken.json";
-import qlfWhiltelistJsonABI from "../artifacts/contracts/WhitelistQlf.sol/WhitelistQlf.json";
-import qlfMaskHolderJsonABI from "../artifacts/contracts/MaskHolderQlf.sol/MaskHolderQlf.json";
-
-import { TestToken } from "../types/contracts/test/test_token.sol/TestToken";
-
-describe("MysteryBox", () => {
-  let signers: Signer[];
-  let contractCreator: Signer;
-  let snapshotId: string;
-
-  let user_1: Signer;
-  let user_2: Signer;
-  let user_3: Signer;
-  let verifier: Signer;
-  let abiCoder: utils.AbiCoder;
-
-  let testTokenAContract: TestToken;
-  let testTokenBContract: TestToken;
-  let testTokenCContract: TestToken;
-  let testMaskTokenContract: TestToken;
-
-  let whitelistQlfContract: Contract;
-  let sigVerifyQlfContract: Contract;
-  let maskHolderQlfContract: Contract;
-  let merkleRootQlfContract: Contract;
-  let mask721ANFTContract: Contract;
-  let mbContract: Contract;
-  let enumerableNftContract: Contract;
-  let nonEnumerableNftContract: Contract;
-
-  // 1 billion tokens, typical decimal 18
-  const testTokenMintAmount = ethers.utils.parseUnits("1000000000", 18).toString();
-  const transferAmount = ethers.utils.parseUnits("100000000", 18).toString();
-
-  before(async () => {
-    signers = await ethers.getSigners();
-    contractCreator = signers[0];
-    // these are whitelisted `box creator`
-    user_1 = signers[1];
-    user_2 = signers[2];
-    user_3 = signers[3];
-    verifier = signers[4];
-    abiCoder = new utils.AbiCoder();
-
-    testTokenAContract = (await deployContract(contractCreator, TestTokenArtifact, [
-      "TestTokenA",
-      "TESTA",
-      testTokenMintAmount,
-    ])) as TestToken;
-    testTokenBContract = (await deployContract(contractCreator, TestTokenArtifact, [
-      "TestTokenB",
-      "TESTB",
-      testTokenMintAmount,
-    ])) as TestToken;
-    testTokenCContract = (await deployContract(contractCreator, TestTokenArtifact, [
-      "TestTokenC",
-      "TESTC",
-      testTokenMintAmount,
-    ])) as TestToken;
-    testMaskTokenContract = (await deployContract(contractCreator, TestTokenArtifact, [
-      "MaskToken",
-      "MASK",
-      testTokenMintAmount,
-    ])) as TestToken;
-
-    {
-      const factory = await ethers.getContractFactory("WhitelistQlf");
-      const proxy = await upgrades.deployProxy(factory, []);
-      whitelistQlfContract = new ethers.Contract(proxy.address, qlfWhiltelistJsonABI.abi, contractCreator);
-    }
-    {
-      const factory = await ethers.getContractFactory("SigVerifyQlf");
-      const SigVerifyQlf = await factory.deploy();
-      sigVerifyQlfContract = await SigVerifyQlf.deployed();
-    }
-    {
-      const factory = await ethers.getContractFactory("MaskHolderQlf");
-      const proxy = await upgrades.deployProxy(factory, [testMaskTokenContract.address, holderMinAmount]);
-      maskHolderQlfContract = new ethers.Contract(proxy.address, qlfMaskHolderJsonABI.abi, contractCreator);
-    }
-    {
-      const factory = await ethers.getContractFactory("MerkleProofQlf");
-      const MerkleProofQlf = await factory.deploy();
-      merkleRootQlfContract = await MerkleProofQlf.deployed();
-    }
-    {
-      const factory = await ethers.getContractFactory("Mask721A");
-      const Mask721A = await factory.deploy(maskNftPara.name, maskNftPara.symbol, maskNftPara.baseURI);
-      mask721ANFTContract = await Mask721A.deployed();
-    }
-    // first is ETH: address(0)
-    createBoxPara.payment.push([testTokenAContract.address, createBoxPara.payment[0][1]]);
-    createBoxPara.payment.push([testTokenBContract.address, createBoxPara.payment[0][1]]);
-    createBoxPara.payment.push([testTokenCContract.address, createBoxPara.payment[0][1]]);
-    {
-      const factory = await ethers.getContractFactory("MaskEnumerableNFT");
-      const proxy = await upgrades.deployProxy(factory, [...Object.values(maskNftPara)]);
-      enumerableNftContract = new ethers.Contract(proxy.address, EnumerableNftTokenABI.abi, contractCreator);
-    }
-    {
-      const factory = await ethers.getContractFactory("MaskNonEnumerableNFT");
-      const proxy = await upgrades.deployProxy(factory, [...Object.values(maskNftPara)]);
-      nonEnumerableNftContract = new ethers.Contract(proxy.address, nonEnumerableNftTokenABI.abi, contractCreator);
-    }
-    {
-      const factory = await ethers.getContractFactory("MysteryBox");
-      const proxy = await upgrades.deployProxy(factory, []);
-      mbContract = new ethers.Contract(proxy.address, jsonABI.abi, contractCreator);
-    }
-    await mbContract.addWhitelist([await user_1.getAddress(), await user_2.getAddress(), await user_3.getAddress()]);
-    await testTokenAContract.transfer(await user_2.getAddress(), transferAmount);
-    await testTokenAContract.connect(user_2).approve(mbContract.address, transferAmount);
-
-    createBoxPara.nft_address = enumerableNftContract.address;
-    // mint 100 NFT for testing
-    await enumerableNftContract.connect(user_1).mint(50);
-    await enumerableNftContract.connect(user_1).mint(50);
-    await enumerableNftContract.connect(user_1).setApprovalForAll(mbContract.address, true);
-
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.sell_all = true;
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      expect(result)
-        .to.have.property("creator")
-        .that.to.be.eq(await user_1.getAddress());
-      expect(result).to.have.property("nft_address").that.to.be.eq(enumerableNftContract.address);
-      expect(result).to.have.property("name").that.to.be.eq(parameter.name);
-      expect(result).to.have.property("start_time").that.to.be.eq(parameter.start_time);
-      expect(result).to.have.property("end_time").that.to.be.eq(parameter.end_time);
-      expect(result).to.have.property("sell_all").that.to.be.eq(parameter.sell_all);
-      expect(result).to.have.property("box_id");
-      sell_all_box_id = result.box_id;
-    }
-    openBoxParameters.box_id = sell_all_box_id;
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.sell_all = false;
-      const nftBalance = await enumerableNftContract.balanceOf(await user_1.getAddress());
-      // half of the NFT ids owned
-      for (let i = 0; i < nftBalance; i += 2) {
-        const nftId = await enumerableNftContract.tokenOfOwnerByIndex(await user_1.getAddress(), i);
-        not_sell_all_nft_id_list.push(nftId);
-      }
-      parameter.nft_id_list = not_sell_all_nft_id_list;
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      expect(result)
-        .to.have.property("creator")
-        .that.to.be.eq(await user_1.getAddress());
-      expect(result).to.have.property("nft_address").that.to.be.eq(enumerableNftContract.address);
-      expect(result).to.have.property("name").that.to.be.eq(parameter.name);
-      expect(result).to.have.property("start_time").that.to.be.eq(parameter.start_time);
-      expect(result).to.have.property("end_time").that.to.be.eq(parameter.end_time);
-      expect(result).to.have.property("sell_all").that.to.be.eq(parameter.sell_all);
-      expect(result).to.have.property("box_id");
-      not_sell_all_box_id = result.box_id;
-
-      const boxStatus = await mbContract.getBoxStatus(not_sell_all_box_id);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(not_sell_all_nft_id_list.length)).to.be.true;
-      expect(boxStatus.total.eq(not_sell_all_nft_id_list.length)).to.be.true;
-      expect(boxStatus).to.have.property("canceled").that.to.be.eq(false);
-
-      const nftList = await mbContract.getNftListForSale(not_sell_all_box_id, 0, parameter.nft_id_list.length);
-      expect(nftList.map((id) => id.toString())).to.eql(parameter.nft_id_list.map((id) => id.toString()));
-    }
-    for (let i = 0; i < 256; i++) {
-      await advanceBlock();
-    }
-  });
-
-  beforeEach(async () => {
-    snapshotId = await takeSnapshot();
-    MockDate.set(Date.now());
-  });
-
-  afterEach(async () => {
-    await revertToSnapShot(snapshotId);
-    // We also need to reset the "real time" (as what we do for evm)
-    MockDate.reset();
-  });
-
-  it("Should variables initialized properly in contract creator", async () => {
-    const owner = await mbContract.owner();
-    expect(owner).to.be.eq(await contractCreator.getAddress());
-  });
-
-  it("Should addNftIntoBox work", async () => {
-    // mint 10 NFT for testing
-    const mintNftAmount = 10;
-    await enumerableNftContract.connect(user_2).mint(mintNftAmount);
-    await enumerableNftContract.connect(user_2).setApprovalForAll(mbContract.address, true);
-    const nftBalance = await enumerableNftContract.balanceOf(await user_2.getAddress());
-    expect(nftBalance.eq(mintNftAmount)).to.be.true;
-
-    let nft_id_list: string[] = [];
-    // half of the NFT ids owned
-    for (let i = 0; i < mintNftAmount / 2; i++) {
-      const nftId = await enumerableNftContract.tokenOfOwnerByIndex(await user_2.getAddress(), i);
-      nft_id_list.push(nftId);
-    }
-    let box_id;
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.sell_all = false;
-      parameter.nft_id_list = nft_id_list;
-      await mbContract.connect(user_2).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      expect(result).to.have.property("box_id");
-      box_id = result.box_id;
-    }
-    {
-      const nftList = await mbContract.getNftListForSale(box_id, 0, mintNftAmount);
-      expect(nftList.map((id) => id.toString())).to.eql(nft_id_list.map((id) => id.toString()));
-
-      const boxStatus = await mbContract.getBoxStatus(box_id);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(nft_id_list.length)).to.be.true;
-      expect(boxStatus.total.eq(nft_id_list.length)).to.be.true;
-      expect(boxStatus).to.have.property("canceled").that.to.be.eq(false);
-    }
-
-    let append_nft_id_list: string[] = [];
-    for (let i = mintNftAmount / 2; i < mintNftAmount; i++) {
-      const nftId = await enumerableNftContract.tokenOfOwnerByIndex(await user_2.getAddress(), i);
-      append_nft_id_list.push(nftId);
-    }
-    await expect(mbContract.connect(user_2).addNftIntoBox(sell_all_box_id, [])).to.be.rejectedWith("not box owner");
-    await expect(mbContract.connect(user_1).addNftIntoBox(sell_all_box_id, [])).to.be.rejectedWith(
-      "can not add for sell_all",
-    );
-    await expect(mbContract.connect(user_1).addNftIntoBox(box_id, append_nft_id_list)).to.be.rejectedWith(
-      "not box owner",
-    );
-
-    {
-      // try adding 'NFT id' he/she does not own
-      let invalid_nft_id: string[] = [];
-      const nftBalance = await enumerableNftContract.balanceOf(await user_1.getAddress());
-      expect(nftBalance.gt(0)).to.be.true;
-      const nftId = await enumerableNftContract.tokenOfOwnerByIndex(await user_1.getAddress(), 0);
-      invalid_nft_id.push(nftId);
-      await expect(mbContract.connect(user_2).addNftIntoBox(box_id, invalid_nft_id)).to.be.rejectedWith(
-        "not nft owner",
-      );
-    }
-    await mbContract.connect(user_2).addNftIntoBox(box_id, append_nft_id_list);
-    {
-      const final_nft_id_list = nft_id_list.concat(append_nft_id_list);
-      const nftList = await mbContract.getNftListForSale(box_id, 0, mintNftAmount);
-      expect(nftList.map((id) => id.toString())).to.eql(final_nft_id_list.map((id) => id.toString()));
-      const boxStatus = await mbContract.getBoxStatus(box_id);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(final_nft_id_list.length)).to.be.true;
-      expect(boxStatus.total.eq(final_nft_id_list.length)).to.be.true;
-      expect(boxStatus).to.have.property("canceled").that.to.be.eq(false);
-    }
-  });
-
-  it("Should cancelBox work", async () => {
-    {
-      const boxStatus = await mbContract.getBoxStatus(sell_all_box_id);
-      expect(boxStatus).to.have.property("started").that.to.be.eq(true);
-      expect(boxStatus).to.have.property("canceled").that.to.be.eq(false);
-      await expect(mbContract.connect(user_2).cancelBox(sell_all_box_id)).to.be.rejectedWith("not box owner");
-      await expect(mbContract.connect(user_1).cancelBox(sell_all_box_id)).to.be.rejectedWith("sale started");
-    }
-    {
-      const open_parameter = JSON.parse(JSON.stringify(openBoxParameters));
-      const now = Math.floor(new Date().getTime() / 1000);
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.start_time = now + seconds_in_a_day;
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      open_parameter.box_id = result.box_id;
-      {
-        const boxStatus = await mbContract.getBoxStatus(open_parameter.box_id);
-        expect(boxStatus).to.have.property("started").that.to.be.eq(false);
-        expect(boxStatus).to.have.property("canceled").that.to.be.eq(false);
-      }
-      await expect(mbContract.connect(user_2).cancelBox(open_parameter.box_id)).to.be.rejectedWith("not box owner");
-      await mbContract.connect(user_1).cancelBox(open_parameter.box_id);
-      await expect(mbContract.connect(user_1).cancelBox(open_parameter.box_id)).to.be.rejectedWith(
-        "sale canceled already",
-      );
-      {
-        const boxStatus = await mbContract.getBoxStatus(open_parameter.box_id);
-        expect(boxStatus).to.have.property("started").that.to.be.eq(false);
-        expect(boxStatus).to.have.property("canceled").that.to.be.eq(true);
-      }
-      await advanceTimeAndBlock(seconds_in_a_day);
-      await expect(
-        mbContract.connect(user_2).openBox(...Object.values(open_parameter), txParameters),
-      ).to.be.rejectedWith("sale canceled");
-    }
-  });
-
-  it("Should getNftListForSale work", async () => {
-    let nft_id_list: string[] = [];
-    const nftBalance = await enumerableNftContract.balanceOf(await user_1.getAddress());
-    // half of the NFT ids owned
-    for (let i = 0; i < nftBalance; i++) {
-      const nftId = await enumerableNftContract.tokenOfOwnerByIndex(await user_1.getAddress(), i);
-      nft_id_list.push(nftId);
-    }
-    const nftList = await mbContract.getNftListForSale(sell_all_box_id, 0, nftBalance);
-    expect(nftList.map((id) => id.toString())).to.eql(nft_id_list.map((id) => id.toString()));
-    {
-      // buy some NFT(s)
-      const parameters = JSON.parse(JSON.stringify(openBoxParameters));
-      parameters.amount = MaxNumberOfNFT;
-      const tx_parameters = {
-        value: txParameters.value.mul(parameters.amount),
-      };
-      await mbContract.connect(user_2).openBox(...Object.values(parameters), tx_parameters);
-    }
-    const newNftList = await mbContract.getNftListForSale(sell_all_box_id, 0, nftBalance);
-    expect(newNftList.length).to.be.eq(nftList.length - MaxNumberOfNFT);
-
-    const purchaseList = await mbContract.getPurchasedNft(sell_all_box_id, await user_2.getAddress());
-    expect(purchaseList.length).to.be.eq(MaxNumberOfNFT);
-    {
-      const nftListStr = nftList.map((id) => id.toString());
-      const newNftListStr = newNftList.map((id) => id.toString());
-      const purchaseListStr = purchaseList.map((id) => id.toString());
-      const mergedList = newNftListStr.concat(purchaseListStr);
-      expect(nftListStr.sort()).to.eql(mergedList.sort());
-    }
-    {
-      const emptyNftList = await mbContract.getNftListForSale(sell_all_box_id, nftBalance, nftBalance);
-      expect(emptyNftList.length).to.be.eq(0);
-    }
-  });
-
-  it("Should not be able to openBox if not-started", async () => {
-    const open_parameter = JSON.parse(JSON.stringify(openBoxParameters));
-    {
-      const now = Math.floor(new Date().getTime() / 1000);
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.start_time = now + seconds_in_a_day;
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      open_parameter.box_id = result.box_id;
-    }
-    {
-      const boxStatus = await mbContract.getBoxStatus(open_parameter.box_id);
-      expect(boxStatus).to.have.property("started").that.to.be.eq(false);
-    }
-    await expect(mbContract.connect(user_2).openBox(...Object.values(open_parameter), txParameters)).to.be.rejectedWith(
-      "not started",
-    );
-  });
-
-  it("Should openBox & getBoxInfo work", async () => {
-    const expectedRemaining = await enumerableNftContract.balanceOf(await user_1.getAddress());
-    const userNftBalance = await enumerableNftContract.balanceOf(await user_2.getAddress());
-    {
-      const boxInfo = await mbContract.getBoxInfo(sell_all_box_id);
-      expect(boxInfo)
-        .to.have.property("creator")
-        .that.to.be.eq(await user_1.getAddress());
-      expect(boxInfo).to.have.property("nft_address").that.to.be.eq(enumerableNftContract.address);
-      expect(boxInfo).to.have.property("name").that.to.be.eq(createBoxPara.name);
-      expect(boxInfo).to.have.property("personal_limit").that.to.be.eq(createBoxPara.personal_limit);
-      expect(boxInfo).to.have.property("qualification").that.to.be.eq(createBoxPara.qualification);
-      expect(boxInfo).to.have.property("qualification_data").that.to.be.eq(createBoxPara.qualification_data);
-
-      const boxStatus = await mbContract.getBoxStatus(sell_all_box_id);
-      expect(boxStatus).to.have.property("payment");
-      expect(boxStatus.payment.map((info) => info.token_addr.toString())).to.eql([
-        createBoxPara.payment[0][0].toString(),
-        createBoxPara.payment[1][0].toString(),
-        createBoxPara.payment[2][0].toString(),
-        createBoxPara.payment[3][0].toString(),
-      ]);
-      expect(boxStatus.payment.map((info) => info.price.toString())).to.eql([
-        createBoxPara.payment[0][1].toString(),
-        createBoxPara.payment[1][1].toString(),
-        createBoxPara.payment[2][1].toString(),
-        createBoxPara.payment[3][1].toString(),
-      ]);
-      expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql(["0", "0", "0", "0"]);
-      expect(boxStatus).to.have.property("started").that.to.be.eq(true);
-      expect(boxStatus).to.have.property("expired").that.to.be.eq(false);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(expectedRemaining)).to.be.true;
-      expect(boxStatus.total.eq(expectedRemaining)).to.be.true;
-    }
-    const tx_parameters = {
-      value: txParameters.value,
-    };
-    await mbContract.connect(user_2).openBox(...Object.values(openBoxParameters), tx_parameters);
-    if (true) {
-      const newUserNftBalance = await enumerableNftContract.balanceOf(await user_2.getAddress());
-      expect(newUserNftBalance.eq(userNftBalance.add(1))).to.be.true;
-      {
-        const logs = await ethers.provider.getLogs(enumerableNftContract.filters.Transfer());
-        const parsedLog = EnumerableNftInterface.parseLog(logs[0]);
-        const result = parsedLog.args;
-        expect(result)
-          .to.have.property("from")
-          .that.to.be.eq(await user_1.getAddress());
-        expect(result)
-          .to.have.property("to")
-          .that.to.be.eq(await user_2.getAddress());
-        expect(result).to.have.property("tokenId");
-      }
-      {
-        const logs = await ethers.provider.getLogs(mbContract.filters.OpenSuccess());
-        const parsedLog = interfaceABI.parseLog(logs[0]);
-        const result = parsedLog.args;
-        expect(result).to.have.property("box_id");
-        expect(result.box_id.eq(openBoxParameters.box_id)).to.be.true;
-        expect(result)
-          .to.have.property("customer")
-          .that.to.be.eq(await user_2.getAddress());
-        expect(result).to.have.property("nft_address").that.to.be.eq(enumerableNftContract.address);
-        expect(result).to.have.property("amount");
-        expect(result.amount.eq(openBoxParameters.amount)).to.be.true;
-      }
-    }
-    {
-      const boxInfo = await mbContract.getBoxInfo(sell_all_box_id);
-      const boxStatus = await mbContract.getBoxStatus(sell_all_box_id);
-      expect(boxInfo)
-        .to.have.property("creator")
-        .that.to.be.eq(await user_1.getAddress());
-      expect(boxInfo).to.have.property("nft_address").that.to.be.eq(enumerableNftContract.address);
-      expect(boxInfo).to.have.property("name").that.to.be.eq(createBoxPara.name);
-      expect(boxInfo).to.have.property("personal_limit").that.to.be.eq(createBoxPara.personal_limit);
-
-      expect(boxStatus).to.have.property("payment");
-      expect(boxStatus.payment.map((info) => info.token_addr.toString())).to.eql([
-        createBoxPara.payment[0][0].toString(),
-        createBoxPara.payment[1][0].toString(),
-        createBoxPara.payment[2][0].toString(),
-        createBoxPara.payment[3][0].toString(),
-      ]);
-      expect(boxStatus.payment.map((info) => info.price.toString())).to.eql([
-        createBoxPara.payment[0][1].toString(),
-        createBoxPara.payment[1][1].toString(),
-        createBoxPara.payment[2][1].toString(),
-        createBoxPara.payment[3][1].toString(),
-      ]);
-      expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql([
-        createBoxPara.payment[0][1].toString(),
-        "0",
-        "0",
-        "0",
-      ]);
-      expect(boxStatus).to.have.property("started").that.to.be.eq(true);
-      expect(boxStatus).to.have.property("expired").that.to.be.eq(false);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(expectedRemaining.sub(1))).to.be.true;
-      expect(boxStatus.total.eq(expectedRemaining)).to.be.true;
-    }
-  });
-
-  it("Should not openBox expired boxes", async () => {
-    const open_parameter = JSON.parse(JSON.stringify(openBoxParameters));
-    {
-      const now = Math.floor(new Date().getTime() / 1000);
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      // to pass `createBox` validation
-      parameter.end_time = now + seconds_in_a_day / 2;
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      open_parameter.box_id = result.box_id;
-    }
-    await advanceTimeAndBlock(seconds_in_a_day);
-    await expect(mbContract.connect(user_2).openBox(...Object.values(open_parameter), txParameters)).to.be.rejectedWith(
-      "expired",
-    );
-    {
-      const boxStatus = await mbContract.getBoxStatus(open_parameter.box_id);
-      expect(boxStatus).to.have.property("expired").that.to.be.eq(true);
-    }
-  });
-
-  it("Should createBox reject invalid parameters", async () => {
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.end_time = 0;
-      await expect(mbContract.connect(user_1).createBox(...Object.values(parameter))).to.be.rejectedWith(
-        "invalid end time",
-      );
-    }
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.payment = [];
-      await expect(mbContract.connect(user_1).createBox(...Object.values(parameter))).to.be.rejectedWith(
-        "invalid payment",
-      );
-    }
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.payment.push([await user_2.getAddress(), createBoxPara.payment[0][1]]);
-      await expect(mbContract.connect(user_1).createBox(...Object.values(parameter))).to.be.rejectedWith(Error);
-    }
-    {
-      await nonEnumerableNftContract.connect(user_1).mint(50);
-      await nonEnumerableNftContract.connect(user_1).setApprovalForAll(mbContract.address, true);
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.nft_address = nonEnumerableNftContract.address;
-      parameter.sell_all = true;
-      await expect(mbContract.connect(user_1).createBox(...Object.values(parameter))).to.be.rejectedWith(
-        "not enumerable nft",
-      );
-    }
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.sell_all = false;
-      parameter.nft_id_list = [];
-      await expect(mbContract.connect(user_1).createBox(...Object.values(parameter))).to.be.rejectedWith(
-        "empty nft list",
-      );
-    }
-    {
-      await enumerableNftContract.connect(user_2).mint(50);
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.sell_all = true;
-      await expect(mbContract.connect(user_2).createBox(...Object.values(parameter))).to.be.rejectedWith(
-        "not ApprovedForAll",
-      );
-    }
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      await enumerableNftContract.connect(user_3).setApprovalForAll(mbContract.address, true);
-      parameter.sell_all = true;
-      await expect(mbContract.connect(user_3).createBox(...Object.values(parameter))).to.be.rejectedWith(
-        "no nft owned",
-      );
-    }
-    {
-      const user_0_NftBalance = await enumerableNftContract.balanceOf(await user_1.getAddress());
-      expect(user_0_NftBalance.gt(0)).to.be.true;
-      const user_1_NftBalance = await enumerableNftContract.balanceOf(await user_2.getAddress());
-      expect(user_1_NftBalance.gt(0)).to.be.true;
-      let nft_id_list: string[] = [];
-      nft_id_list.push(await enumerableNftContract.tokenOfOwnerByIndex(await user_1.getAddress(), 0));
-      nft_id_list.push(await enumerableNftContract.tokenOfOwnerByIndex(await user_2.getAddress(), 0));
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.sell_all = false;
-      parameter.nft_id_list = nft_id_list;
-      await expect(mbContract.connect(user_1).createBox(...Object.values(parameter))).to.be.rejectedWith("now owner");
-    }
-  });
-
-  it("Should openBox reject invalid parameters", async () => {
-    {
-      const parameters = JSON.parse(JSON.stringify(openBoxParameters));
-      parameters.payment_token_index = 1;
-      await testTokenAContract.connect(user_3).approve(mbContract.address, transferAmount);
-      await expect(mbContract.connect(user_3).openBox(...Object.values(parameters))).to.be.rejectedWith(
-        "ERC20: transfer amount exceeds balance",
-      );
-      await testTokenAContract.connect(user_3).approve(mbContract.address, 0);
-      await testTokenAContract.transfer(await user_3.getAddress(), transferAmount);
-      await expect(mbContract.connect(user_3).openBox(...Object.values(parameters))).to.be.rejectedWith(
-        "ERC20: insufficient allowance",
-      );
-    }
-    {
-      const parameters = JSON.parse(JSON.stringify(openBoxParameters));
-      parameters.amount = MaxNumberOfNFT + 1;
-      await expect(mbContract.connect(user_3).openBox(...Object.values(parameters))).to.be.rejectedWith(
-        "exceeds personal limit",
-      );
-    }
-    {
-      const parameters = JSON.parse(JSON.stringify(openBoxParameters));
-      parameters.payment_token_index = createBoxPara.payment.length;
-      await expect(mbContract.connect(user_3).openBox(...Object.values(parameters))).to.be.rejectedWith(
-        "invalid payment token",
-      );
-    }
-    {
-      const tx_parameters = {
-        value: 0,
-      };
-      await expect(
-        mbContract.connect(user_3).openBox(...Object.values(openBoxParameters), tx_parameters),
-      ).to.be.rejectedWith("not enough ETH");
-    }
-  });
-
-  it("Should personal limit work", async () => {
-    const nftBalance = await enumerableNftContract.balanceOf(await user_1.getAddress());
-    const contractEthBalanceBeforeOpen = await ethers.provider.getBalance(mbContract.address);
-    const userEthBalanceBeforeOpen = await ethers.provider.getBalance(await user_2.getAddress());
-    const parameters = JSON.parse(JSON.stringify(openBoxParameters));
-    parameters.amount = MaxNumberOfNFT / 2;
-    const tx_parameters = {
-      value: txParameters.value.mul(parameters.amount),
-    };
-    await mbContract.connect(user_2).openBox(...Object.values(parameters), tx_parameters);
-    {
-      const paymentEth = tx_parameters.value;
-      const contractEthBalanceAfterOpen = await ethers.provider.getBalance(mbContract.address);
-      const userEthBalanceAfterOpen = await ethers.provider.getBalance(await user_2.getAddress());
-      expect(contractEthBalanceAfterOpen.eq(contractEthBalanceBeforeOpen.add(paymentEth))).to.be.true;
-      expect(userEthBalanceBeforeOpen.gt(userEthBalanceAfterOpen.add(paymentEth))).to.be.true;
-    }
-    await mbContract.connect(user_2).openBox(...Object.values(parameters), tx_parameters);
-    {
-      const paymentEth = tx_parameters.value.mul(2);
-      const contractEthBalanceAfterOpen = await ethers.provider.getBalance(mbContract.address);
-      const userEthBalanceAfterOpen = await ethers.provider.getBalance(await user_2.getAddress());
-      expect(contractEthBalanceAfterOpen.eq(contractEthBalanceBeforeOpen.add(paymentEth))).to.be.true;
-      expect(userEthBalanceBeforeOpen.gt(userEthBalanceAfterOpen.add(paymentEth))).to.be.true;
-    }
-    await expect(mbContract.connect(user_2).openBox(...Object.values(parameters), tx_parameters)).to.be.rejectedWith(
-      "exceeds personal limit",
-    );
-    {
-      const boxInfo = await mbContract.getBoxInfo(parameters.box_id);
-      const boxStatus = await mbContract.getBoxStatus(parameters.box_id);
-      expect(boxInfo)
-        .to.have.property("creator")
-        .that.to.be.eq(await user_1.getAddress());
-      expect(boxInfo).to.have.property("nft_address").that.to.be.eq(enumerableNftContract.address);
-      expect(boxInfo).to.have.property("name").that.to.be.eq(createBoxPara.name);
-      expect(boxStatus).to.have.property("started").that.to.be.eq(true);
-      expect(boxStatus).to.have.property("expired").that.to.be.eq(false);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(nftBalance.sub(MaxNumberOfNFT))).to.be.true;
-      expect(boxStatus.total.eq(nftBalance)).to.be.true;
-    }
-  });
-
-  it('Should sell "sell_all_box_id" NFT work', async () => {
-    const totalNFT = await enumerableNftContract.balanceOf(await user_1.getAddress());
-    expect((await enumerableNftContract.balanceOf(await signers[5].getAddress())).eq(0)).to.be.true;
-    expect((await enumerableNftContract.balanceOf(await signers[6].getAddress())).eq(0)).to.be.true;
-    expect((await enumerableNftContract.balanceOf(await signers[7].getAddress())).eq(0)).to.be.true;
-    expect((await enumerableNftContract.balanceOf(await signers[8].getAddress())).eq(0)).to.be.true;
-    expect((await enumerableNftContract.balanceOf(await signers[9].getAddress())).eq(0)).to.be.true;
-
-    const contractEthBalanceBeforeOpen = await ethers.provider.getBalance(mbContract.address);
-    const user_8_EthBalance = await ethers.provider.getBalance(await signers[8].getAddress());
-    const user_9_EthBalance = await ethers.provider.getBalance(await signers[9].getAddress());
-    const open_parameters = JSON.parse(JSON.stringify(openBoxParameters));
-    {
-      open_parameters.amount = MaxNumberOfNFT;
-      const tx_parameters = {
-        value: txParameters.value.mul(open_parameters.amount),
-      };
-      await mbContract.connect(signers[5]).openBox(...Object.values(open_parameters), tx_parameters);
-      await mbContract.connect(signers[6]).openBox(...Object.values(open_parameters), tx_parameters);
-      await mbContract.connect(signers[7]).openBox(...Object.values(open_parameters), tx_parameters);
-      await mbContract.connect(signers[8]).openBox(...Object.values(open_parameters), tx_parameters);
-      await expect(mbContract.connect(user_1).claimPayment([open_parameters.box_id])).to.be.rejectedWith(
-        "not expired/sold-out",
-      );
-      await mbContract.connect(signers[9]).openBox(...Object.values(open_parameters), tx_parameters);
-      await expect(
-        mbContract.connect(signers[10]).openBox(...Object.values(open_parameters), tx_parameters),
-      ).to.be.rejectedWith("no NFT left");
-      expect((await enumerableNftContract.balanceOf(await signers[5].getAddress())).eq(MaxNumberOfNFT)).to.be.true;
-      expect((await enumerableNftContract.balanceOf(await signers[6].getAddress())).eq(MaxNumberOfNFT)).to.be.true;
-      expect((await enumerableNftContract.balanceOf(await signers[7].getAddress())).eq(MaxNumberOfNFT)).to.be.true;
-      expect((await enumerableNftContract.balanceOf(await signers[8].getAddress())).eq(MaxNumberOfNFT)).to.be.true;
-      const remainingNFT = totalNFT.sub(MaxNumberOfNFT * 4);
-      expect((await enumerableNftContract.balanceOf(await signers[9].getAddress())).eq(remainingNFT)).to.be.true;
-
-      const user_8_EthBalanceAfter = await ethers.provider.getBalance(await signers[8].getAddress());
-      const user_9_EthBalanceAfter = await ethers.provider.getBalance(await signers[9].getAddress());
-      expect(user_8_EthBalance.gt(user_8_EthBalanceAfter.add(tx_parameters.value))).to.be.true;
-      // test ETH refund
-      expect(user_9_EthBalanceAfter.gt(user_9_EthBalance.sub(tx_parameters.value))).to.be.true;
-    }
-    const contractEthBalanceAfterOpen = await ethers.provider.getBalance(mbContract.address);
-    const totalPaymentEth = txParameters.value.mul(totalNFT);
-    expect(contractEthBalanceAfterOpen.eq(contractEthBalanceBeforeOpen.add(totalPaymentEth))).to.be.true;
-    {
-      const boxInfo = await mbContract.getBoxInfo(sell_all_box_id);
-      const boxStatus = await mbContract.getBoxStatus(sell_all_box_id);
-      expect(boxInfo)
-        .to.have.property("creator")
-        .that.to.be.eq(await user_1.getAddress());
-      expect(boxInfo).to.have.property("nft_address").that.to.be.eq(enumerableNftContract.address);
-      expect(boxInfo).to.have.property("name").that.to.be.eq(createBoxPara.name);
-      expect(boxInfo).to.have.property("personal_limit").that.to.be.eq(createBoxPara.personal_limit);
-
-      expect(boxStatus).to.have.property("payment");
-      expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql([
-        totalPaymentEth.toString(),
-        "0",
-        "0",
-        "0",
-      ]);
-      expect(boxStatus).to.have.property("started").that.to.be.eq(true);
-      expect(boxStatus).to.have.property("expired").that.to.be.eq(false);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(0)).to.be.true;
-      expect(boxStatus.total.eq(totalNFT)).to.be.true;
-    }
-    await mbContract.connect(user_1).claimPayment([open_parameters.box_id]);
-    {
-      const boxStatus = await mbContract.getBoxStatus(sell_all_box_id);
-      expect(boxStatus).to.have.property("payment");
-      expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql(["0", "0", "0", "0"]);
-    }
-  });
-
-  it("Should support lots of NFT", async () => {
-    for (let i = 0; i < 18; i++) {
-      await enumerableNftContract.connect(user_1).mint(50);
-    }
-    let id_list: string[] = [];
-    let box_id;
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.personal_limit = 255;
-      parameter.sell_all = false;
-      const nftBalance = await enumerableNftContract.balanceOf(await user_1.getAddress());
-      // all the NFT ids owned, total 1000 NFT(s)
-      for (let i = 0; i < nftBalance; i++) {
-        const nftId = await enumerableNftContract.tokenOfOwnerByIndex(await user_1.getAddress(), i);
-        id_list.push(nftId);
-      }
-      parameter.nft_id_list = id_list;
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      expect(result)
-        .to.have.property("creator")
-        .that.to.be.eq(await user_1.getAddress());
-      expect(result).to.have.property("nft_address").that.to.be.eq(enumerableNftContract.address);
-      expect(result).to.have.property("name").that.to.be.eq(parameter.name);
-      expect(result).to.have.property("start_time").that.to.be.eq(parameter.start_time);
-      expect(result).to.have.property("end_time").that.to.be.eq(parameter.end_time);
-      expect(result).to.have.property("sell_all").that.to.be.eq(parameter.sell_all);
-      expect(result).to.have.property("box_id");
-      box_id = result.box_id;
-
-      const boxStatus = await mbContract.getBoxStatus(box_id);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(id_list.length)).to.be.true;
-      expect(boxStatus.total.eq(id_list.length)).to.be.true;
-      expect(boxStatus).to.have.property("canceled").that.to.be.eq(false);
-
-      const nftList = await mbContract.getNftListForSale(box_id, 0, parameter.nft_id_list.length);
-      expect(nftList.map((id) => id.toString())).to.eql(parameter.nft_id_list.map((id) => id.toString()));
-    }
-    // mint another 1000 NFT(s)
-    for (let i = 0; i < 20; i++) {
-      await enumerableNftContract.connect(user_1).mint(50);
-    }
-    let id_list_batch_2: string[] = [];
-    {
-      const nftBalance = await enumerableNftContract.balanceOf(await user_1.getAddress());
-      // all the NFT ids owned, total 1000 NFT(s)
-      for (let i = id_list.length; i < nftBalance; i++) {
-        const nftId = await enumerableNftContract.tokenOfOwnerByIndex(await user_1.getAddress(), i);
-        id_list_batch_2.push(nftId);
-      }
-      await mbContract.connect(user_1).addNftIntoBox(box_id, id_list_batch_2);
-      const boxStatus = await mbContract.getBoxStatus(box_id);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(id_list.length + id_list_batch_2.length)).to.be.true;
-      expect(boxStatus.total.eq(id_list.length + id_list_batch_2.length)).to.be.true;
-      expect(boxStatus).to.have.property("canceled").that.to.be.eq(false);
-
-      const nftList = await mbContract.getNftListForSale(box_id, id_list.length, id_list_batch_2.length);
-      expect(nftList.map((id) => id.toString())).to.eql(id_list_batch_2.map((id) => id.toString()));
-    }
-    const nftBalance_before_batch_3 = await enumerableNftContract.balanceOf(await user_1.getAddress());
-    // mint another 1000 NFT(s)
-    for (let i = 0; i < 20; i++) {
-      await enumerableNftContract.connect(user_1).mint(50);
-    }
-    let id_list_batch_3: string[] = [];
-    {
-      const nftBalance = await enumerableNftContract.balanceOf(await user_1.getAddress());
-      // all the NFT ids owned, total 1000 NFT(s)
-      for (let i = nftBalance_before_batch_3; i < nftBalance; i++) {
-        const nftId = await enumerableNftContract.tokenOfOwnerByIndex(await user_1.getAddress(), i);
-        id_list_batch_3.push(nftId);
-      }
-      await mbContract.connect(user_1).addNftIntoBox(box_id, id_list_batch_3);
-      const boxStatus = await mbContract.getBoxStatus(box_id);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(nftBalance_before_batch_3.add(id_list_batch_3.length))).to.be.true;
-      expect(boxStatus.total.eq(nftBalance_before_batch_3.add(id_list_batch_3.length))).to.be.true;
-      expect(boxStatus).to.have.property("canceled").that.to.be.eq(false);
-
-      const nftList = await mbContract.getNftListForSale(box_id, nftBalance_before_batch_3, id_list_batch_3.length);
-      expect(nftList.map((id) => id.toString())).to.eql(id_list_batch_3.map((id) => id.toString()));
-    }
-    {
-      const nftBalance_before = await enumerableNftContract.balanceOf(await user_1.getAddress());
-      const buy_batch = 255;
-      // buy 255 NFT(s)
-      const parameters = JSON.parse(JSON.stringify(openBoxParameters));
-      parameters.box_id = box_id;
-      parameters.amount = buy_batch;
-      const tx_parameters = {
-        value: txParameters.value.mul(parameters.amount),
-      };
-      await mbContract.connect(user_2).openBox(...Object.values(parameters), tx_parameters);
-      const user_2_balance = await enumerableNftContract.balanceOf(await user_2.getAddress());
-      const nftBalance_after = await enumerableNftContract.balanceOf(await user_1.getAddress());
-
-      expect(user_2_balance.eq(255)).to.be.true;
-      const boxStatus = await mbContract.getBoxStatus(box_id);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(nftBalance_before.sub(buy_batch))).to.be.true;
-      expect(boxStatus.remaining.eq(nftBalance_after)).to.be.true;
-    }
-  });
-
-  it('Should sell "not_sell_all" enumerableNftContract work', async () => {
-    const boxStatus = await mbContract.getBoxStatus(not_sell_all_box_id);
-    expect(boxStatus).to.have.property("remaining");
-    const totalNFT = boxStatus.remaining;
-    {
-      await testTokenAContract.transfer(await signers[5].getAddress(), transferAmount);
-      await testTokenAContract.connect(signers[5]).approve(mbContract.address, transferAmount);
-      await testTokenAContract.transfer(await signers[6].getAddress(), transferAmount);
-      await testTokenAContract.connect(signers[6]).approve(mbContract.address, transferAmount);
-      await testTokenAContract.transfer(await signers[7].getAddress(), transferAmount);
-      await testTokenAContract.connect(signers[7]).approve(mbContract.address, transferAmount);
-      await testTokenAContract.transfer(await signers[8].getAddress(), transferAmount);
-      await testTokenAContract.connect(signers[8]).approve(mbContract.address, transferAmount);
-    }
-    const contractTokenBalanceBefore = await testTokenAContract.balanceOf(mbContract.address);
-    {
-      expect((await enumerableNftContract.balanceOf(await signers[5].getAddress())).eq(0)).to.be.true;
-      expect((await enumerableNftContract.balanceOf(await signers[6].getAddress())).eq(0)).to.be.true;
-      expect((await enumerableNftContract.balanceOf(await signers[7].getAddress())).eq(0)).to.be.true;
-      const parameters = JSON.parse(JSON.stringify(openBoxParameters));
-      parameters.amount = MaxNumberOfNFT;
-      parameters.payment_token_index = 1;
-      parameters.box_id = not_sell_all_box_id;
-      await mbContract.connect(signers[5]).openBox(...Object.values(parameters));
-      await mbContract.connect(signers[6]).openBox(...Object.values(parameters));
-      await mbContract.connect(signers[7]).openBox(...Object.values(parameters));
-      await expect(mbContract.connect(signers[8]).openBox(...Object.values(parameters))).to.be.rejectedWith(
-        "no NFT left",
-      );
-      expect((await enumerableNftContract.balanceOf(await signers[5].getAddress())).eq(MaxNumberOfNFT)).to.be.true;
-      expect((await enumerableNftContract.balanceOf(await signers[6].getAddress())).eq(MaxNumberOfNFT)).to.be.true;
-      const remainingNFT = totalNFT.sub(MaxNumberOfNFT * 2);
-      expect((await enumerableNftContract.balanceOf(await signers[7].getAddress())).eq(remainingNFT)).to.be.true;
-    }
-    // sold out, validate payment token amount
-    const paymentTokenAAmount = createBoxPara.payment[0][1].mul(totalNFT);
-    const contractTokenBalanceAfter = await testTokenAContract.balanceOf(mbContract.address);
-    expect(contractTokenBalanceAfter.eq(contractTokenBalanceBefore.add(paymentTokenAAmount))).to.be.true;
-    {
-      const boxStatus = await mbContract.getBoxStatus(not_sell_all_box_id);
-      expect(boxStatus).to.have.property("payment");
-      expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql([
-        "0",
-        paymentTokenAAmount.toString(),
-        "0",
-        "0",
-      ]);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(0)).to.be.true;
-      expect(boxStatus.total.eq(not_sell_all_nft_id_list.length)).to.be.true;
-    }
-  });
-
-  it('Should sell "not_sell_all" nonEnumerableNftContract work', async () => {
-    let box_id;
-    const total = 50;
-    {
-      await nonEnumerableNftContract.connect(user_1).mint(total);
-      await nonEnumerableNftContract.connect(user_1).setApprovalForAll(mbContract.address, true);
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.nft_address = nonEnumerableNftContract.address;
-      parameter.sell_all = false;
-      const nftBalance = await nonEnumerableNftContract.balanceOf(await user_1.getAddress());
-      expect(total.toString()).to.be.eq(nftBalance.toString());
-
-      let id_list: number[] = [];
-      for (let i = 0; i < nftBalance; i++) {
-        id_list.push(i);
-      }
-      parameter.nft_id_list = id_list;
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      expect(result)
-        .to.have.property("creator")
-        .that.to.be.eq(await user_1.getAddress());
-      expect(result).to.have.property("nft_address").that.to.be.eq(nonEnumerableNftContract.address);
-      expect(result).to.have.property("name").that.to.be.eq(parameter.name);
-      expect(result).to.have.property("start_time").that.to.be.eq(parameter.start_time);
-      expect(result).to.have.property("end_time").that.to.be.eq(parameter.end_time);
-      expect(result).to.have.property("sell_all").that.to.be.eq(parameter.sell_all);
-      expect(result).to.have.property("box_id");
-      box_id = result.box_id;
-    }
-
-    const boxStatus = await mbContract.getBoxStatus(box_id);
-    expect(boxStatus).to.have.property("remaining");
-    expect(total.toString()).to.be.eq(boxStatus.remaining.toString());
-    const totalNFT = boxStatus.remaining;
-    {
-      await testTokenAContract.transfer(await signers[5].getAddress(), transferAmount);
-      await testTokenAContract.connect(signers[5]).approve(mbContract.address, transferAmount);
-      await testTokenAContract.transfer(await signers[6].getAddress(), transferAmount);
-      await testTokenAContract.connect(signers[6]).approve(mbContract.address, transferAmount);
-      await testTokenAContract.transfer(await signers[7].getAddress(), transferAmount);
-      await testTokenAContract.connect(signers[7]).approve(mbContract.address, transferAmount);
-      await testTokenAContract.transfer(await signers[8].getAddress(), transferAmount);
-      await testTokenAContract.connect(signers[8]).approve(mbContract.address, transferAmount);
-    }
-    const contractTokenBalanceBefore = await testTokenAContract.balanceOf(mbContract.address);
-    {
-      expect((await nonEnumerableNftContract.balanceOf(await signers[5].getAddress())).eq(0)).to.be.true;
-      expect((await nonEnumerableNftContract.balanceOf(await signers[6].getAddress())).eq(0)).to.be.true;
-      expect((await nonEnumerableNftContract.balanceOf(await signers[7].getAddress())).eq(0)).to.be.true;
-      const parameters = JSON.parse(JSON.stringify(openBoxParameters));
-      parameters.amount = MaxNumberOfNFT;
-      parameters.payment_token_index = 1;
-      parameters.box_id = box_id;
-      await mbContract.connect(signers[5]).openBox(...Object.values(parameters));
-      await mbContract.connect(signers[6]).openBox(...Object.values(parameters));
-      await mbContract.connect(signers[7]).openBox(...Object.values(parameters));
-      await expect(mbContract.connect(signers[8]).openBox(...Object.values(parameters))).to.be.rejectedWith(
-        "no NFT left",
-      );
-      expect((await nonEnumerableNftContract.balanceOf(await signers[5].getAddress())).eq(MaxNumberOfNFT)).to.be.true;
-      expect((await nonEnumerableNftContract.balanceOf(await signers[6].getAddress())).eq(MaxNumberOfNFT)).to.be.true;
-      const remainingNFT = totalNFT.sub(MaxNumberOfNFT * 2);
-      expect((await nonEnumerableNftContract.balanceOf(await signers[7].getAddress())).eq(remainingNFT)).to.be.true;
-    }
-    // sold out, validate payment token amount
-    const paymentTokenAAmount = createBoxPara.payment[0][1].mul(totalNFT);
-    const contractTokenBalanceAfter = await testTokenAContract.balanceOf(mbContract.address);
-    expect(contractTokenBalanceAfter.eq(contractTokenBalanceBefore.add(paymentTokenAAmount))).to.be.true;
-    {
-      const boxStatus = await mbContract.getBoxStatus(box_id);
-      expect(boxStatus).to.have.property("payment");
-      expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql([
-        "0",
-        paymentTokenAAmount.toString(),
-        "0",
-        "0",
-      ]);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(0)).to.be.true;
-      expect(boxStatus.total.eq(totalNFT)).to.be.true;
-    }
-  });
-
-  it('Should work if "NFT tokens transfered elsewhere"', async () => {
-    const boxStatus = await mbContract.getBoxStatus(not_sell_all_box_id);
-    expect(boxStatus).to.have.property("remaining");
-    expect(boxStatus.remaining.eq(not_sell_all_nft_id_list.length)).to.be.true;
-
-    // transfer NFT(s) elsewhere, remaining random numebr of NFT(s)
-    const max = MaxNumberOfNFT / 2;
-    const remain = Math.floor(Math.random() * max);
-    // console.log(remain);
-    for (let i = remain; i < not_sell_all_nft_id_list.length; i++) {
-      await enumerableNftContract
-        .connect(user_1)
-        .transferFrom(await user_1.getAddress(), await user_2.getAddress(), not_sell_all_nft_id_list[i]);
-    }
-    const testAccount = signers[5];
-    await testTokenAContract.transfer(await testAccount.getAddress(), transferAmount);
-    await testTokenAContract.connect(testAccount).approve(mbContract.address, transferAmount);
-    const parameters = JSON.parse(JSON.stringify(openBoxParameters));
-    parameters.amount = max;
-    parameters.payment_token_index = 1;
-    parameters.box_id = not_sell_all_box_id;
-
-    while (true) {
-      const boxStatus = await mbContract.getBoxStatus(not_sell_all_box_id);
-      if (boxStatus.remaining.eq(0)) {
-        break;
-      }
-
-      await mbContract.connect(testAccount).openBox(...Object.values(parameters));
-      {
-        const logs = await ethers.provider.getLogs(mbContract.filters.OpenSuccess());
-        const parsedLog = interfaceABI.parseLog(logs[0]);
-        const result = parsedLog.args;
-        expect(result).to.have.property("amount");
-        // console.log("amount: " + result.amount.toString());
-      }
-    }
-    expect((await enumerableNftContract.balanceOf(await testAccount.getAddress())).eq(remain)).to.be.true;
-  });
-
-  it("Should 721A NFT work", async () => {
-    let box_id;
-    const contractEthBalanceBeforeOpen = await ethers.provider.getBalance(mbContract.address);
-    const userEthBalanceBeforeOpen = await ethers.provider.getBalance(await user_2.getAddress());
-    const batch = await mask721ANFTContract.BATCH_SIZE();
-    await mask721ANFTContract.connect(user_1).mint(batch);
-    const totalNFT = await mask721ANFTContract.balanceOf(await user_1.getAddress());
-    expect(totalNFT.eq(batch)).to.be.true;
-    expect((await mask721ANFTContract.balanceOf(await signers[5].getAddress())).eq(0)).to.be.true;
-    await mask721ANFTContract.connect(user_1).setApprovalForAll(mbContract.address, true);
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.nft_address = mask721ANFTContract.address;
-      parameter.sell_all = true;
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      expect(result)
-        .to.have.property("creator")
-        .that.to.be.eq(await user_1.getAddress());
-      expect(result).to.have.property("nft_address").that.to.be.eq(mask721ANFTContract.address);
-      expect(result).to.have.property("name").that.to.be.eq(parameter.name);
-      expect(result).to.have.property("start_time").that.to.be.eq(parameter.start_time);
-      expect(result).to.have.property("end_time").that.to.be.eq(parameter.end_time);
-      expect(result).to.have.property("sell_all").that.to.be.eq(parameter.sell_all);
-      expect(result).to.have.property("box_id");
-      box_id = result.box_id;
-    }
-    const parameters = JSON.parse(JSON.stringify(openBoxParameters));
-    parameters.box_id = box_id;
-    parameters.amount = batch;
-    const tx_parameters = {
-      value: txParameters.value.mul(parameters.amount),
-    };
-    await mbContract.connect(user_2).openBox(...Object.values(parameters), tx_parameters);
-    {
-      const paymentEth = tx_parameters.value;
-      const contractEthBalanceAfterOpen = await ethers.provider.getBalance(mbContract.address);
-      const userEthBalanceAfterOpen = await ethers.provider.getBalance(await user_2.getAddress());
-      expect(contractEthBalanceAfterOpen.eq(contractEthBalanceBeforeOpen.add(paymentEth))).to.be.true;
-      expect(userEthBalanceBeforeOpen.gt(userEthBalanceAfterOpen.add(paymentEth))).to.be.true;
-      const user_nft_balance = await mask721ANFTContract.balanceOf(await user_2.getAddress());
-      expect(totalNFT.eq(user_nft_balance)).to.be.true;
-      const owner_nft_balance = await mask721ANFTContract.balanceOf(await user_1.getAddress());
-      expect(owner_nft_balance.eq(0)).to.be.true;
-    }
-    await expect(mbContract.connect(user_2).openBox(...Object.values(parameters), tx_parameters)).to.be.rejectedWith(
-      "no NFT left",
-    );
-    {
-      const boxInfo = await mbContract.getBoxInfo(parameters.box_id);
-      const boxStatus = await mbContract.getBoxStatus(parameters.box_id);
-      expect(boxInfo)
-        .to.have.property("creator")
-        .that.to.be.eq(await user_1.getAddress());
-      expect(boxInfo).to.have.property("nft_address").that.to.be.eq(mask721ANFTContract.address);
-      expect(boxInfo).to.have.property("name").that.to.be.eq(createBoxPara.name);
-      expect(boxStatus).to.have.property("started").that.to.be.eq(true);
-      expect(boxStatus).to.have.property("expired").that.to.be.eq(false);
-      expect(boxStatus).to.have.property("remaining");
-      expect(boxStatus.remaining.eq(0)).to.be.true;
-      expect(boxStatus.total.eq(totalNFT)).to.be.true;
-    }
-  });
-
-  it("Should claimPayment reject invalid parameters", async () => {
-    await expect(mbContract.connect(user_2).claimPayment([sell_all_box_id])).to.be.rejectedWith("not owner");
-    await expect(mbContract.connect(user_2).claimPayment([not_sell_all_box_id])).to.be.rejectedWith("not owner");
-  });
-
-  it("Should claimPayment work", async () => {
-    {
-      const boxStatus = await mbContract.getBoxStatus(not_sell_all_box_id);
-      expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql(["0", "0", "0", "0"]);
-    }
-    const tx_parameters = {
-      value: txParameters.value,
-    };
-    const open_parameters = JSON.parse(JSON.stringify(openBoxParameters));
-    open_parameters.box_id = not_sell_all_box_id;
-    await mbContract.connect(user_2).openBox(...Object.values(open_parameters), tx_parameters);
-    const paymentEthAmount = createBoxPara.payment[0][1];
-    const paymentTokenAAmount = createBoxPara.payment[1][1];
-    {
-      const boxStatus = await mbContract.getBoxStatus(not_sell_all_box_id);
-      expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql([
-        paymentEthAmount.toString(),
-        "0",
-        "0",
-        "0",
-      ]);
-    }
-    open_parameters.payment_token_index = 1;
-    await mbContract.connect(user_2).openBox(...Object.values(open_parameters));
-    {
-      const boxStatus = await mbContract.getBoxStatus(not_sell_all_box_id);
-      expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql([
-        paymentEthAmount.toString(),
-        paymentTokenAAmount.toString(),
-        "0",
-        "0",
-      ]);
-    }
-    {
-      const sell_all_open_parameters = JSON.parse(JSON.stringify(openBoxParameters));
-      sell_all_open_parameters.box_id = sell_all_box_id;
-      await mbContract.connect(user_2).openBox(...Object.values(sell_all_open_parameters), tx_parameters);
-      {
-        const boxStatus = await mbContract.getBoxStatus(not_sell_all_box_id);
-        expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql([
-          paymentEthAmount.toString(),
-          paymentTokenAAmount.toString(),
-          "0",
-          "0",
-        ]);
-      }
-      {
-        const boxStatus = await mbContract.getBoxStatus(sell_all_box_id);
-        expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql([
-          paymentEthAmount.toString(),
-          "0",
-          "0",
-          "0",
-        ]);
-      }
-    }
-    // ETH balance before
-    const authorEthBalanceBeforeClaim = await ethers.provider.getBalance(await user_1.getAddress());
-    const contractEthBalanceBeforeClaim = await ethers.provider.getBalance(mbContract.address);
-    // token A balance before
-    const authorTokenABalanceBeforeClaim = await testTokenAContract.balanceOf(await user_1.getAddress());
-    const contractTokenABalanceBeforeClaim = await testTokenAContract.balanceOf(mbContract.address);
-    await expect(mbContract.connect(user_1).claimPayment([not_sell_all_box_id])).to.be.rejectedWith(
-      "not expired/sold-out",
-    );
-    await expect(mbContract.connect(user_2).claimPayment([not_sell_all_box_id])).to.be.rejectedWith("not owner");
-
-    await advanceTimeAndBlock(seconds_in_a_day * 90);
-    await mbContract.connect(user_1).claimPayment([not_sell_all_box_id]);
-    {
-      const boxStatus = await mbContract.getBoxStatus(not_sell_all_box_id);
-      // `claimPayment` can only claim payment in one of the boxes
-      expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql(["0", "0", "0", "0"]);
-    }
-    const authorEthBalanceAfterClaim = await ethers.provider.getBalance(await user_1.getAddress());
-    const contractEthBalanceAfterClaim = await ethers.provider.getBalance(mbContract.address);
-    // token A balance after
-    const authorTokenABalanceAfterClaim = await testTokenAContract.balanceOf(await user_1.getAddress());
-    const contractTokenABalanceAfterClaim = await testTokenAContract.balanceOf(mbContract.address);
-    // assume `gas consumption` is less than `paymentEthAmount`
-    expect(authorEthBalanceAfterClaim.gt(authorEthBalanceBeforeClaim)).to.be.true;
-    expect(contractEthBalanceBeforeClaim.eq(contractEthBalanceAfterClaim.add(paymentEthAmount))).to.be.true;
-    expect(authorTokenABalanceAfterClaim.eq(authorTokenABalanceBeforeClaim.add(paymentTokenAAmount))).to.be.true;
-    expect(contractTokenABalanceBeforeClaim.eq(contractTokenABalanceAfterClaim.add(paymentTokenAAmount))).to.be.true;
-    {
-      const boxStatus = await mbContract.getBoxStatus(sell_all_box_id);
-      // `claimPayment` can only claim payment in one of the boxes
-      expect(boxStatus.payment.map((info) => info.receivable_amount.toString())).to.eql([
-        paymentEthAmount.toString(),
-        "0",
-        "0",
-        "0",
-      ]);
-    }
-    await mbContract.connect(user_1).claimPayment([not_sell_all_box_id]);
-    {
-      const authorEthBalanceAfterClaimAgain = await ethers.provider.getBalance(await user_1.getAddress());
-      const contractEthBalanceAfterClaimAgain = await ethers.provider.getBalance(mbContract.address);
-      expect(authorEthBalanceAfterClaim.gt(authorEthBalanceAfterClaimAgain)).to.be.true;
-      expect(contractEthBalanceAfterClaim.eq(contractEthBalanceAfterClaimAgain)).to.be.true;
-
-      // token A balance after
-      const authorTokenABalanceAfterClaimAgain = await testTokenAContract.balanceOf(await user_1.getAddress());
-      const contractTokenABalanceAfterClaimAgain = await testTokenAContract.balanceOf(mbContract.address);
-      expect(authorTokenABalanceAfterClaim.eq(authorTokenABalanceAfterClaimAgain)).to.be.true;
-      expect(contractTokenABalanceAfterClaim.eq(contractTokenABalanceAfterClaimAgain)).to.be.true;
-    }
-    {
-      const contractEthBalance = await ethers.provider.getBalance(mbContract.address);
-      expect(contractEthBalance.gt(0)).to.be.true;
-    }
-  });
-
-  it("Should MaskHolder work", async () => {
-    expect((await testMaskTokenContract.balanceOf(await contractCreator.getAddress())).gt(0)).to.be.true;
-    expect((await testMaskTokenContract.balanceOf(await user_1.getAddress())).eq(0)).to.be.true;
-
-    const open_parameters = JSON.parse(JSON.stringify(openBoxParameters));
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.holder_token_addr = testMaskTokenContract.address;
-      parameter.sell_all = true;
-      parameter.holder_min_token_amount = holderMinAmount;
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      open_parameters.box_id = result.box_id;
-    }
-    await expect(
-      mbContract.connect(user_1).openBox(...Object.values(open_parameters), txParameters),
-    ).to.be.rejectedWith("not holding enough token");
-
-    // transfer
-    await testMaskTokenContract.transfer(await user_1.getAddress(), holderMinAmount);
-    await mbContract.connect(user_1).openBox(...Object.values(open_parameters), txParameters);
-
-    const boxInfo = await mbContract.getBoxInfo(open_parameters.box_id);
-    expect(boxInfo).to.have.property("holder_min_token_amount");
-    expect(boxInfo.holder_min_token_amount.eq(holderMinAmount)).to.be.true;
-    expect(boxInfo).to.have.property("holder_token_addr");
-    expect(boxInfo.holder_token_addr).to.be.eq(testMaskTokenContract.address);
-  });
-
-  it("Should addAdmin/addWhitelist work", async () => {
-    let user_test = signers[4];
-    await enumerableNftContract.connect(user_test).mint(10);
-    await enumerableNftContract.connect(user_test).setApprovalForAll(mbContract.address, true);
-    const parameter = JSON.parse(JSON.stringify(createBoxPara));
-    parameter.sell_all = true;
-
-    expect(await mbContract.whitelist(await user_1.getAddress())).to.be.eq(true);
-    expect(await mbContract.admin(await user_1.getAddress())).to.be.eq(false);
-    await expect(mbContract.connect(user_1).addAdmin([await user_test.getAddress()])).to.be.rejectedWith(
-      "Ownable: caller is not the owner",
-    );
-    await expect(mbContract.connect(user_1).addWhitelist([await user_test.getAddress()])).to.be.rejectedWith(
-      "not admin",
-    );
-    await expect(mbContract.connect(user_test).createBox(...Object.values(parameter))).to.be.rejectedWith(
-      "not whitelisted",
-    );
-    expect(await mbContract.whitelist(await user_test.getAddress())).to.be.eq(false);
-    expect(await mbContract.admin(await user_test.getAddress())).to.be.eq(false);
-    await mbContract.connect(contractCreator).addAdmin([await user_1.getAddress()]);
-    expect(await mbContract.admin(await user_1.getAddress())).to.be.eq(true);
-    await mbContract.connect(user_1).addWhitelist([await user_test.getAddress()]);
-    expect(await mbContract.whitelist(await user_test.getAddress())).to.be.eq(true);
-    await expect(mbContract.connect(user_1).addAdmin([await user_test.getAddress()])).to.be.rejectedWith(
-      "Ownable: caller is not the owner",
-    );
-    await mbContract.connect(user_test).createBox(...Object.values(parameter));
-
-    await expect(mbContract.connect(user_test).removeWhitelist([await user_test.getAddress()])).to.be.rejectedWith(
-      "not admin",
-    );
-    await mbContract.connect(user_1).removeWhitelist([await user_test.getAddress()]);
-    expect(await mbContract.whitelist(await user_test.getAddress())).to.be.eq(false);
-    await expect(mbContract.connect(user_test).createBox(...Object.values(parameter))).to.be.rejectedWith(
-      "not whitelisted",
-    );
-
-    await mbContract.connect(contractCreator).removeAdmin([await user_1.getAddress()]);
-    expect(await mbContract.admin(await user_1.getAddress())).to.be.eq(false);
-    await expect(mbContract.connect(user_1).addWhitelist([await user_test.getAddress()])).to.be.rejectedWith(
-      "not admin",
-    );
-  });
-
-  it("Should whitelist qualification work", async () => {
-    const open_parameters = JSON.parse(JSON.stringify(openBoxParameters));
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.sell_all = true;
-      parameter.qualification = whitelistQlfContract.address;
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      open_parameters.box_id = result.box_id;
-    }
-    await whitelistQlfContract.addWhitelist([await user_1.getAddress()]);
-    await mbContract.connect(user_1).openBox(...Object.values(open_parameters), txParameters);
-    await expect(
-      mbContract.connect(user_2).openBox(...Object.values(open_parameters), txParameters),
-    ).to.be.rejectedWith("not whitelisted");
-
-    const boxInfo = await mbContract.getBoxInfo(open_parameters.box_id);
-    expect(boxInfo).to.have.property("qualification").that.to.be.eq(whitelistQlfContract.address);
-  });
-
-  it("Should signature verification qualification work", async () => {
-    const open_parameters = JSON.parse(JSON.stringify(openBoxParameters));
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.sell_all = true;
-      parameter.qualification = sigVerifyQlfContract.address;
-      // convert address to bytes32
-      const qualification_data = utils.hexZeroPad((await verifier.getAddress()).toString(), 32);
-      parameter.qualification_data = qualification_data;
-
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      open_parameters.box_id = result.box_id;
-    }
-    {
-      const msg_hash = utils.keccak256(await user_1.getAddress());
-      const signature = await verifier.signMessage(ethers.utils.arrayify(msg_hash));
-      open_parameters.proof = signature;
-      await mbContract.connect(user_1).openBox(...Object.values(open_parameters), txParameters);
-      await expect(
-        mbContract.connect(user_2).openBox(...Object.values(open_parameters), txParameters),
-      ).to.be.rejectedWith("not qualified");
-    }
-    const boxInfo = await mbContract.getBoxInfo(open_parameters.box_id);
-    expect(boxInfo).to.have.property("qualification").that.to.be.eq(sigVerifyQlfContract.address);
-  });
-
-  it("Should merkle root qualification work", async () => {
-    const open_parameters = JSON.parse(JSON.stringify(openBoxParameters));
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.sell_all = true;
-      parameter.qualification = merkleRootQlfContract.address;
-      parameter.qualification_data = proofs.merkleRoot;
-
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      open_parameters.box_id = result.box_id;
-    }
-    {
-      const boxInfo = await mbContract.getBoxInfo(open_parameters.box_id);
-      expect(boxInfo).to.have.property("qualification").that.to.be.eq(merkleRootQlfContract.address);
-      expect(boxInfo).to.have.property("qualification_data").that.to.be.eq(proofs.merkleRoot);
-    }
-    {
-      let user_1_index = proofs.leaves.length;
-      for (let i = 0; i < proofs.leaves.length; i++) {
-        if (proofs.leaves[i].address === (await user_1.getAddress())) {
-          user_1_index = i;
-        }
-      }
-      expect(user_1_index < proofs.leaves.length).to.be.true;
-      const proof = abiCoder.encode(["bytes32[]"], [proofs.leaves[user_1_index].proof]);
-      open_parameters.proof = proof;
-      await mbContract.connect(user_1).openBox(...Object.values(open_parameters), txParameters);
-      await expect(
-        mbContract.connect(user_2).openBox(...Object.values(open_parameters), txParameters),
-      ).to.be.rejectedWith("not qualified");
-    }
-    {
-      let user_2_index = proofs.leaves.length;
-      for (let i = 0; i < proofs.leaves.length; i++) {
-        if (proofs.leaves[i].address === (await user_2.getAddress())) {
-          user_2_index = i;
-        }
-      }
-      expect(user_2_index < proofs.leaves.length).to.be.true;
-
-      const proof = abiCoder.encode(["bytes32[]"], [proofs.leaves[user_2_index].proof]);
-      open_parameters.proof = proof;
-      await expect(
-        mbContract.connect(user_2).setQualificationData(open_parameters.box_id, createBoxPara.qualification_data),
-      ).to.be.rejectedWith("not box owner");
-      await mbContract.connect(user_1).setQualificationData(open_parameters.box_id, createBoxPara.qualification_data);
-      {
-        const boxInfo = await mbContract.getBoxInfo(open_parameters.box_id);
-        expect(boxInfo).to.have.property("qualification_data").that.to.be.eq(createBoxPara.qualification_data);
-      }
-      await expect(
-        mbContract.connect(user_2).openBox(...Object.values(open_parameters), txParameters),
-      ).to.be.rejectedWith("not qualified");
-      await mbContract.connect(user_1).setQualificationData(open_parameters.box_id, proofs.merkleRoot);
-      {
-        const boxInfo = await mbContract.getBoxInfo(open_parameters.box_id);
-        expect(boxInfo).to.have.property("qualification_data").that.to.be.eq(proofs.merkleRoot);
-      }
-      await mbContract.connect(user_2).openBox(...Object.values(open_parameters), txParameters);
-    }
-  });
-
-  it("Should MaskHolderQlf qualification work", async () => {
-    expect((await testMaskTokenContract.balanceOf(await contractCreator.getAddress())).gt(0)).to.be.true;
-    expect((await testMaskTokenContract.balanceOf(await user_1.getAddress())).eq(0)).to.be.true;
-
-    const open_parameters = JSON.parse(JSON.stringify(openBoxParameters));
-    {
-      const parameter = JSON.parse(JSON.stringify(createBoxPara));
-      parameter.sell_all = true;
-      parameter.qualification = maskHolderQlfContract.address;
-      await mbContract.connect(user_1).createBox(...Object.values(parameter));
-      const logs = await ethers.provider.getLogs(mbContract.filters.CreationSuccess());
-      const parsedLog = interfaceABI.parseLog(logs[0]);
-      const result = parsedLog.args;
-      open_parameters.box_id = result.box_id;
-    }
-    await expect(
-      mbContract.connect(user_1).openBox(...Object.values(open_parameters), txParameters),
-    ).to.be.rejectedWith("not holding enough token");
-
-    // transfer
-    await testMaskTokenContract.transfer(await user_1.getAddress(), holderMinAmount);
-    await mbContract.connect(user_1).openBox(...Object.values(open_parameters), txParameters);
-
-    const boxInfo = await mbContract.getBoxInfo(open_parameters.box_id);
-    expect(boxInfo).to.have.property("qualification").that.to.be.eq(maskHolderQlfContract.address);
-  });
-
-  it("Should not be able to call initialize", async () => {
-    await expect(mbContract.connect(contractCreator).initialize()).to.be.rejectedWith(Error);
-    await expect(mbContract.connect(user_1).initialize()).to.be.rejectedWith(Error);
-    await expect(mbContract.connect(user_2).initialize()).to.be.rejectedWith(Error);
-  });
-
-  it("Should transferOwnership work", async () => {
-    await expect(mbContract.connect(user_1).transferOwnership(await user_1.getAddress())).to.be.rejectedWith(
-      "Ownable: caller is not the owner",
-    );
-    await mbContract.connect(contractCreator).transferOwnership(await user_1.getAddress());
-    const owner = await mbContract.owner();
-    expect(owner).to.be.eq(await user_1.getAddress());
-  });
-});
